Temperature Sensor and graphing tool
This is a program that will connect to an arduino through USB and create a graph of the temperature recorded by an included temperature sensor. A function is also included that will write values to .csv instead of graphing.
For this to work, you will need:
    - An arduino or similar
    - An LM35 temperature sensor wired to port A0
This project was intended to be a proportional-integral temperature controller, but the arrival of the transistors I need for my heating element has been greatly delayed, so for now it will just be a temperature sensor.